colors management for all widgets fixed, some cleanup
