--- graphs/src/cz/nkp/urnnbn/client/charts/widgets/LifecycleHandler.java
@@ -1,6 +1,0 @@
-package cz.nkp.urnnbn.client.charts.widgets;
-
-public interface LifecycleHandler {
-
-	public void onAfterDrawn();
-}
