+++ graphs/src/cz/nkp/urnnbn/client/charts/widgets/topLevel/ColorConstants.java
@@ -2,20 +2,23 @@ package cz.nkp.urnnbn.client.charts.widgets.topLevel;
 
 public class ColorConstants {
 
+	// http://paletton.com/#uid=7030u0kw0vSjzD3oSy0y9oLDhjs
+	// primary-0
 	public static final String ASSIGNMENTS_VALUE_NEUTRAL = "#FF6F63";
+	// complement-2
+	public static final String ASSIGNMENTS_VALUE_1 = "#00C222";
+	// secondery2-2
+	public static final String ASSIGNMENTS_VALUE_2 = "#03899C";
+	// secondery-2
+	public static final String ASSIGNMENTS_VALUE_3 = "#FE7A00";
 
+	// http://paletton.com/#uid=73k0X0kHRr1r0CGEE-YLClVQkgi
+	// primary-0
 	public static final String RESOLVATIONS_VALUE_NEUTRAL = "#1EA6B3";
-
-	// case URN_NBN_ASSIGNMENTS:
-	// // http://paletton.com/#uid=7030u0kw0vSjzD3oSy0y9oLDhjs
-	// // primary-0, complement-2, secondery2-2, secondery-2
-	// return new String[] { "#FF6F63", "#00C222", "#03899C", "#FE7A00" };
-	// case URN_NBN_RESOLVATIONS:
-	// // http://paletton.com/#uid=73k0X0kHRr1r0CGEE-YLClVQkgi
-	// // primary-0, complement-2, secondery2-2, secondery-2
-	// return new String[] { "#1EA6B3", "#D76500", "#D7B500", "#2E0A94" };
-	// default:
-	// return null;
-	// }
-
+	// complement-2
+	public static final String RESOLVATIONS_VALUE_1 = "#D76500";
+	// secondery2-2
+	public static final String RESOLVATIONS_VALUE_2 = "#D7B500";
+	// secondery-2
+	public static final String RESOLVATIONS_VALUE_3 = "#2E0A94";
 }
